Add script for linking library files into bin directory.
For some reason sqlplus (and probably other binaries tools) are not using $DYLD_LIBRARY_PATH and returning an error complaining about libraries not being loaded.

The easiest fix would be to dump everything into the bin directory like the official install script usually does. But then what's the point of this "traditional" structure install?

This script will link all libraries into the bin directory. The bin directory will still be full of links but at least we keep the structure and environment variables intact and it's easy to go back to the usual way as soon as this get fixed by Oracle.